Reset event store recorded events collection before triggering post commit event
When the event store triggers the post commit event all the transaction
related aspect should be closed, even the collection of recorded events
stored during the transaction and made persistent into the evento store
should be empty.
This because if the listener of the post commit event wants to start a
new transaction it should not take care of the reset ot the recorded
events collection. It's not its resposibility.
If the listener is insterested into what events has been committed it
can look into the event itself using the getRecordeEvents method.